CHET-400: Fix missing imports after package rename
