+++ core/src/main/java/com/atlassian/migration/datacenter/core/fs/captor/AttachmentCaptor.java
@@ -16,8 +16,6 @@
 
 package com.atlassian.migration.datacenter.core.fs.captor;
 
-import com.atlassian.jira.issue.attachment.Attachment;
-
 import java.nio.file.Path;
 
 @FunctionalInterface
