+++ jira-plugin/src/main/java/com/atlassian/migration/datacenter/configuration/MigrationAssistantProfileSpecificConfiguration.java
@@ -18,8 +18,8 @@ import com.atlassian.jira.config.util.JiraHome;
 import com.atlassian.jira.issue.attachment.AttachmentStore;
 import com.atlassian.migration.datacenter.core.application.ApplicationConfiguration;
 import com.atlassian.migration.datacenter.core.aws.AllowAnyTransitionMigrationServiceFacade;
-import com.atlassian.migration.datacenter.core.fs.capture.AttachmentCaptor;
-import com.atlassian.migration.datacenter.core.fs.capture.JiraIssueAttachmentListener;
+import com.atlassian.migration.datacenter.core.fs.captor.AttachmentCaptor;
+import com.atlassian.migration.datacenter.core.fs.captor.JiraIssueAttachmentListener;
 import com.atlassian.migration.datacenter.spi.MigrationService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
